Fixed few breaking tests
