+++ test/EIP20CoGateway/confirmStakeIntent.js
@@ -25,6 +25,7 @@ const sinon = require('sinon');
 const assert = chai.assert;
 const EIP20CoGateway = require('../../src/ContractInteract/EIP20CoGateway');
 const SpyAssert = require('../../test_utils/SpyAssert');
+const AssertAsync = require('../../test_utils/AssertAsync');
 const Utils = require('../../src/utils/Utils');
 
 describe('EIP20CoGateway.confirmStakeIntent()', () => {
@@ -52,7 +53,7 @@ describe('EIP20CoGateway.confirmStakeIntent()', () => {
     spySendTransaction = sinon.replace(
       Utils,
       'sendTransaction',
-      sinon.fake.returns(true),
+      sinon.fake.resolves(true),
     );
   };
 
@@ -89,8 +90,8 @@ describe('EIP20CoGateway.confirmStakeIntent()', () => {
   });
 
   it('should throw error transaction object is invalid', async () => {
-    assert.throws(() => {
-      coGateway.confirmStakeIntent(
+    AssertAsync.throws(async () => {
+      await coGateway.confirmStakeIntent(
         stakeParams.staker,
         stakeParams.nonce,
         stakeParams.beneficiary,
