+++ test/EIP20Gateway/proveGateway.js
@@ -25,6 +25,7 @@ const sinon = require('sinon');
 const assert = chai.assert;
 const EIP20Gateway = require('../../src/ContractInteract/EIP20Gateway');
 const SpyAssert = require('../../test_utils/SpyAssert');
+const AssertAsync = require('../../test_utils/AssertAsync');
 const Utils = require('../../src/utils/Utils');
 
 describe('EIP20Gateway.proveGateway()', () => {
@@ -55,7 +56,7 @@ describe('EIP20Gateway.proveGateway()', () => {
     spySendTransaction = sinon.replace(
       Utils,
       'sendTransaction',
-      sinon.fake.returns(true),
+      sinon.fake.resolves(true),
     );
   };
 
@@ -85,8 +86,8 @@ describe('EIP20Gateway.proveGateway()', () => {
   });
 
   it('should throw error transaction object is invalid', async () => {
-    assert.throws(() => {
-      gateway.proveGateway(
+    AssertAsync.throws(async () => {
+      await gateway.proveGateway(
         blockHeight,
         encodedAccount,
         accountProof,
