--- test/stake_helper/getCoGatewayStateRootProvider.js
@@ -1,97 +1,0 @@
-// Copyright 2019 OpenST Ltd.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-//
-// ----------------------------------------------------------------------------
-//
-// http://www.simpletoken.org/
-//
-// ----------------------------------------------------------------------------
-
-const chai = require('chai');
-const sinon = require('sinon');
-const Web3 = require('web3');
-const StakeHelper = require('../../src/helpers/StakeHelper');
-const Contracts = require('../../src/Contracts');
-const SpyAssert = require('../../test_utils/SpyAssert');
-
-const assert = chai.assert;
-
-describe('StakeHelper.getCoGatewayStateRootProvider()', () => {
-  let stakeHelper;
-  let originWeb3;
-  let auxiliaryWeb3;
-  let gatewayAddress;
-  let coGatewayAddress;
-
-  beforeEach(() => {
-    // runs before each test in this block
-    originWeb3 = new Web3('http://localhost:8545');
-    auxiliaryWeb3 = new Web3('http://localhost:8546');
-    gatewayAddress = '0x0000000000000000000000000000000000000001';
-    coGatewayAddress = '0x0000000000000000000000000000000000000002';
-    stakeHelper = new StakeHelper(
-      originWeb3,
-      auxiliaryWeb3,
-      gatewayAddress,
-      coGatewayAddress,
-    );
-  });
-
-  it('should return correct state root provider contract address', async function() {
-    const expectedAddress = '0x0000000000000000000000000000000000000003';
-
-    // Mock an instance of cogateway contract.
-    const mockContract = sinon.mock(
-      Contracts.getEIP20CoGateway(auxiliaryWeb3, coGatewayAddress),
-    );
-    const coGatewayContract = mockContract.object;
-
-    // Fake the stateRootProvider call.
-    const spyStateRootProvider = sinon.replace(
-      coGatewayContract.methods,
-      'stateRootProvider',
-      sinon.fake.returns(() => {
-        return Promise.resolve(expectedAddress);
-      }),
-    );
-
-    const spyContract = sinon.replace(
-      Contracts,
-      'getEIP20CoGateway',
-      sinon.fake.returns(coGatewayContract),
-    );
-
-    // Add spy on stakeHelper.getCoGatewayStateRootProvider.
-    const spy = sinon.spy(stakeHelper, 'getCoGatewayStateRootProvider');
-
-    // Get auxiliary chain's state root provider contract address.
-    const stateRootProviderAddress = await stakeHelper.getCoGatewayStateRootProvider();
-
-    // Assert the returned value.
-    assert.strictEqual(
-      stateRootProviderAddress,
-      expectedAddress,
-      'State root provider contract address must not be different.',
-    );
-
-    SpyAssert.assert(spyStateRootProvider, 1, [[]]);
-    SpyAssert.assert(spyContract, 1, [[auxiliaryWeb3, coGatewayAddress]]);
-    SpyAssert.assert(spy, 1, [[]]);
-
-    // Restore all mocked and spy objects.
-    mockContract.restore();
-    spy.restore();
-    sinon.restore();
-  });
-});
