+++ tests/test_anchor.js
